Move jit to cuda/
