+++ ffcx_backends/cuda/__init__.py
@@ -1,9 +1,11 @@
 """Custom CUDA FFCx backend."""
 
 from ffcx_backends.cuda import expression, file, form, integral
-from ffcx_backends.cuda.jit import compile_objects
+from ffcx_backends.cuda.jit import compile_expressions, compile_forms
 
 __all__ = [
+    "compile_expressions",
+    "compile_forms",
     "compile_objects",
     "expression",
     "file",
